Create CardIDs
This is the beginning of using separate card ids and then assigning suit and value based off of the card id. I also return the entire list that contains the deck so that I can modify entries on the list and hopefully this will give me the ability to keep score and decrease from the deck so that I don't play the same card over and over. It's not quite working yet but it has some good code that I can tweak and I think that it is a valuable branch to go off of